refactor: merge controllers into a single file
